Refactored tests and updated imports
