+++ flash/text/__init__.py
@@ -1,7 +1,6 @@
 from flash.text.classification import TextClassificationData, TextClassifier  # noqa: F401
+from flash.text.question_answering import QuestionAnsweringData, QuestionAnsweringTask  # noqa: F401
 from flash.text.seq2seq import (  # noqa: F401
-    QuestionAnsweringData,
-    QuestionAnsweringTask,
     Seq2SeqData,
     Seq2SeqTask,
     SummarizationData,
