+++ flash/text/question_answering/__init__.py
@@ -1,2 +1,2 @@
-from flash.text.seq2seq.question_answering.data import QuestionAnsweringData  # noqa: F401
-from flash.text.seq2seq.question_answering.model import QuestionAnsweringTask  # noqa: F401
+from flash.text.question_answering.data import QuestionAnsweringData  # noqa: F401
+from flash.text.question_answering.model import QuestionAnsweringTask  # noqa: F401
